Auth/requirements.txt
Django==5.2.3
python-decouple
google-generativeai
